py-cython: updated to 0.28.2
0.28.2 (2018-04-13)
Features added
abs() is faster for Python long objects.
The C++11 methods front() and end() were added to the declaration of libcpp.string.
The C++11 methods reserve() and bucket_count() are declared for libcpp.unordered_map.

Bugs fixed
The copy of a read-only memoryview was considered read-only as well, whereas a common reason to copy a read-only view is to make it writable. The result of the copying is now a writable buffer by default.
The switch statement generation failed to apply recursively to the body of converted if-statements.
NULL was sometimes rejected as exception return value when the returned type is a fused pointer type.
Fixed compatibility with PyPy 5.11.

Other changes
The NumPy tutorial was rewritten to use memoryviews instead of the older buffer declaration syntax.